add a hollowed-out implementation of the i2c peripheral code.
As-per Scott's recommendation on how to get started, this is basically just the atmel-samd header and empty function signatures.  it compiles, and I've tested that I
can load it onto a real circ-py board and then "import i2cperipheral" (from the REPL).  Just putting this in as a place-holder and a starting point.  Presumably we
now just need to populate these 9 functions.